+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
   
+## 0.1.0
+
+  * Added `Live` function
+  * Updated `emit` event function to be called `trigger`
+  * Added `src` dir to bower distributable
+
 ## 0.0.1
 
   * Created